move servlet req/resp wrappers to parent package
